test: enable QuickOpen-integ-test
